// src/common/enums.ts
export enum ObjectIcons {
  RELOAD = "refresh",
  PASTE = "content_paste",
  CUT = "content_cut",
  COPY = "content_copy",
  DELETE = "clear",
  RENAME = "edit_note",
  NEW_FOLDER = "create_new_folder",
  NEW_OBJECT = "note_add",
  NEW_LABEL = "new_label",
  OPEN = "auto_stories",
  EDIT = "mode_edit",
  SELECT_ALL = "toc",
  DESELECT_ALL = "list",
  SEARCH = "search",
  PROPERTIES = "rate_review",

  TEST = "test",
}